+++ src/engine/_edit.rs
@@ -4,7 +4,7 @@ mod _holds;
 mod _trnsp;
 mod _vfc_sets;
 
-use crate::*;
+use crate::engine::*;
 use better_default::Default;
 
 #[derive(Default, Clone)]
