+++ src/engine/_edit/_dig_inputs.rs
@@ -1,4 +1,4 @@
-use crate::*;
+use crate::engine::*;
 use duplicate::duplicate_item;
 use std::cmp::Ordering::{Equal, Greater, Less};
 use winit::keyboard::KeyCode;
