+++ src/engine/_edit/_trnsp.rs
@@ -1,4 +1,4 @@
-use crate::{engine::_edit::{NewTrait, NewEnginePartParams}, *};
+use crate::engine::{_edit::*, *};
 use duplicate::duplicate_item;
 
 #[duplicate_item(
